+++ src/App/Storages/SDCard.cpp
@@ -1,4 +1,4 @@
-#include <App/SDCard.h>
+#include <App/Storages/SDCard.h>
 
 namespace App {
 
@@ -158,7 +158,7 @@ namespace App {
       const char *tShort = GetFileName(tFileEntry.name());
       if (IsFile(tShort)) {
         char tBuffer[16];
-        UTL.ByteToReadableSize(tFileEntry.size(), tBuffer, sizeof(tBuffer));
+        UTL.ByteToReadableSize((uint64_t)tFileEntry.size(), tBuffer, sizeof(tBuffer));
         AppendToBuffer(tShort, strlen(tShort));
         AppendToBuffer(" [", 2);
         AppendToBuffer(tBuffer, strlen(tBuffer));
@@ -431,7 +431,7 @@ namespace App {
     if (tSize > sizeof(mFileBuffer)) {
       tFile.close();
       char tSizeBuffer[16];
-      UTL.ByteToReadableSize((uint32_t)tSize, tSizeBuffer, sizeof(tSizeBuffer));
+      UTL.ByteToReadableSize((uint64_t)tSize, tSizeBuffer, sizeof(tSizeBuffer));
       snprintf(mFileBuffer, sizeof(mFileBuffer), "  Error: File too large (%s).\r\n", tSizeBuffer);
       return mFileBuffer;
     }
